chore/PRSD-NONE update IncompletePropertiesViewModel to use ViewModelBuilder pattern
